Clear external and resident data when persisting the process

+++ helpers/persistProcessData.ts
@@ -1,6 +1,8 @@
 import { NextRouter } from "next/router";
 import { TransactionMode } from "remultiform/database";
+import { externalStoreNames } from "../storage/ExternalDatabaseSchema";
 import { processStoreNames } from "../storage/ProcessDatabaseSchema";
+import { residentStoreNames } from "../storage/ResidentDatabaseSchema";
 import Storage from "../storage/Storage";
 import basePath from "./basePath";
 import getProcessApiJwt from "./getProcessApiJwt";
@@ -16,58 +18,48 @@ const persistProcessData = async (
     setProgress(progress);
   }
 
-  if (Storage.ProcessContext && Storage.ProcessContext.database) {
-    const processRef = getProcessRef(router);
-    const processApiJwt = getProcessApiJwt(processRef);
+  if (
+    !Storage.ExternalContext ||
+    !Storage.ExternalContext.database ||
+    !Storage.ResidentContext ||
+    !Storage.ResidentContext.database ||
+    !Storage.ProcessContext ||
+    !Storage.ProcessContext.database
+  ) {
+    return;
+  }
 
-    if (!processRef || !processApiJwt) {
-      console.error(
-        "Unable to persist process data due to missing session data"
-      );
+  const processRef = getProcessRef(router);
+  const processApiJwt = getProcessApiJwt(processRef);
 
-      return;
-    }
+  if (!processRef || !processApiJwt) {
+    console.error("Unable to persist process data due to missing session data");
 
-    const json = await Storage.getProcessJson(processRef);
+    return;
+  }
 
-    if (json) {
-      const { processJson, imagesJson } = json;
+  const json = await Storage.getProcessJson(processRef);
 
-      const progressIncrement = 1 / (imagesJson.length + 2);
+  if (!json) {
+    console.warn("No process data to persist");
 
-      await Promise.all(
-        imagesJson.map(async ({ id, image }) => {
-          const response = await fetch(
-            `${basePath}/api/v1/processes/${processRef}/images?jwt=${processApiJwt}`,
-            {
-              method: "POST",
-              headers: {
-                "Content-Type": "application/json",
-              },
-              body: JSON.stringify({ id, image }),
-            }
-          );
+    return;
+  }
 
-          if (!response.ok) {
-            throw new Error(`${response.status}: ${response.statusText}`);
-          }
+  const { processJson, imagesJson } = json;
 
-          progress += progressIncrement;
+  const progressIncrement = 1 / (imagesJson.length + 2);
 
-          if (setProgress) {
-            setProgress(progress);
-          }
-        })
-      );
-
+  await Promise.all(
+    imagesJson.map(async ({ id, image }) => {
       const response = await fetch(
-        `${basePath}/api/v1/processes/${processRef}/processData?jwt=${processApiJwt}`,
+        `${basePath}/api/v1/processes/${processRef}/images?jwt=${processApiJwt}`,
         {
-          method: "PATCH",
+          method: "POST",
           headers: {
             "Content-Type": "application/json",
           },
-          body: JSON.stringify(processJson),
+          body: JSON.stringify({ id, image }),
         }
       );
 
@@ -80,29 +72,83 @@ const persistProcessData = async (
       if (setProgress) {
         setProgress(progress);
       }
+    })
+  );
 
-      const db = await Storage.ProcessContext.database;
+  const response = await fetch(
+    `${basePath}/api/v1/processes/${processRef}/processData?jwt=${processApiJwt}`,
+    {
+      method: "PATCH",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(processJson),
+    }
+  );
 
-      // To reduce risk of data loss, we only clear up the data if we sent
-      // something to the backend.
-      await db.transaction(
-        processStoreNames,
+  if (!response.ok) {
+    throw new Error(`${response.status}: ${response.statusText}`);
+  }
+
+  progress += progressIncrement;
+
+  if (setProgress) {
+    setProgress(progress);
+  }
+
+  // To reduce risk of data loss, we only clear up the data if we sent
+  // something to the backend.
+  {
+    const externalDatabase = await Storage.ExternalContext.database;
+    const residentDatabase = await Storage.ResidentContext.database;
+    const processDatabase = await Storage.ProcessContext.database;
+
+    await externalDatabase.transaction(
+      externalStoreNames,
+      async (stores) => {
+        await Promise.all(
+          Object.values(stores).map((store) => store.delete(processRef))
+        );
+      },
+      TransactionMode.ReadWrite
+    );
+
+    const residentRefs = await processDatabase.get(
+      "tenantsPresent",
+      processRef
+    );
+
+    if (residentRefs) {
+      await residentDatabase.transaction(
+        residentStoreNames,
         async (stores) => {
           await Promise.all(
-            Object.values(stores).map((store) => store.delete(processRef))
+            Object.values(stores).map(async (store) => {
+              for (const ref of residentRefs) {
+                await store.delete(ref);
+              }
+            })
           );
         },
         TransactionMode.ReadWrite
       );
-
-      sessionStorage.clear();
-
-      if (setProgress) {
-        setProgress(1);
-      }
     }
-  } else {
-    console.warn("No process data to persist");
+
+    await processDatabase.transaction(
+      processStoreNames,
+      async (stores) => {
+        await Promise.all(
+          Object.values(stores).map((store) => store.delete(processRef))
+        );
+      },
+      TransactionMode.ReadWrite
+    );
+
+    sessionStorage.clear();
+  }
+
+  if (setProgress) {
+    setProgress(1);
   }
 };
